requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
crispy-bootstrap5==2024.10
cryptography==43.0.3
cssselect2==0.7.0
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.1.2
django-allauth==65.0.2
django-autocomplete-light==3.11.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-summernote==0.8.20.0
environs==11.0.0
html5lib==1.1
idna==3.10
inflection==0.5.1
lxml==5.3.0
marshmallow==3.23.0
oscrypto==1.3.0
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
